Numerical modular symbols. Initial commit.
This is the initial commit to move the previous steps to git. So there is quite a lot of things here.
This implements a new version of modular symbols. The main faile is mod_sym.pyx. The rest are adaption so that it can be used. This new implementation is very different from eclib and sage's native modular symbols. It computes a precise numerical approximation to the integral of the modular form along paths in the upper half plane. It should be faster when the conductor is large compared to how many symbols need to be computed.