Calling submatrix() method in smith_form() with nrows and ncols defined explicitly.
Sometimes the submatrix() method does not determine the nrows and ncols of the resultant submatrix correctly (such as when calculating the smith normal form of a matrix mod 2). This explicit statement fixes that. All the example code still works as commented in the source, but now we can compute the integers mod 2 smith normal form