More thematic tutorials and demos
This patch adds a collection of tutorials and demos that have been
used during various Sage(-Combinat) Days.

 - #8886: object and class demo for sage days 20.5
 - Sage-combinat tutorial
 - ...

Rationale for writing such tutorial and demos in ReST rather than
using worksheets:

 - version control
 - use of standard editing tools
 - doctesting
 - use of the same ReST syntax as for the rest of the documentation.
   this makes it easy to transfer chunks to/from it

Rationale for (eventually) centralizing all tutorials in the Sage
distribution (or temporarily in the Sage-Combinat queue):

 - ease to find for users and speakers
 - fosters reusing and shared ownership
 - ease of distribution of a large number of worksheets (see #10652)

Rationale for putting the tutorials directly in the source tree:

 - keep them close to the source
   fosters consistency, doctesting, ...
 - keep related demo/primer/tutorial all together, naturally grouped by theme
 - allows for ``sage.combinat.tutorial?`` in the Sage interpreter / notebook

Cons: there might not be a natural place for transverse tutorials

Rationale for using a docstring within a .py file:

 - sage -t does not currently behave exactly the same way on .rst
   files and .py files (loss of context between chunks, ...)
 - currently required for sage.combinat.tutorial? to work
 - no need to add a label on top of each file to link to it

Cons:
 - Is it possible to use includes to build a single worksheet out of
   several pieces?
 - Need to use an import in __init__ (one more thing to do; may use resources)

Those are mostly technical issues that could possibly be fixed.

Further issues:
 - support for .. MODULEAUTHOR:: ... in a .py?
 - (see e-mail to Florent)