Faster exactification using numeric minpoly.
An algebraic number which was constructed from a symbolic expression will
now maintain a reference to said symbolic expression.  If exactification is
needed, the numeric method for computing a minimal polynomial of the
expression will be tried, which will likely yield a better result than
recursive traversal of the algebraic number descriptors that resulted from
the conversion of individual subexpressions.