DirichletGroups over rings with unhashable elements
DirichletGroups construct a map from zeta^i to i. This is not possible if the
powers of zeta are not hashable. This happens for some p-adics (cf. ticket
11895) and should probably be the case in other inexact rings. One could
imagine having a different structure than a hash map for computing discrete
logs quickly in such cases (e.g. a prefix-tree which can be traversed by
digits). This commit disables the map when it did not work anyway and replaces
it with a trivial implementation in such cases.