Added a pickle parameter for @cached_method
This can be used to control whether a cache should be pickled or not. The
default is not to pickle caches. Often caches are not very expensive to
recompute, they are just to expensive to compute very often. Additionally
caching causes trouble in some places. It tends to lead to pickling problems
due to circular references in combination with instances of
CachedRepresentation (#15155). It is also dangerous in combination with special
methods such as __hash__ (#15270).