Trac #15006: Added clarification about lack of order in dictionaries.
The last output does not really illustrate that dictionaries do not have order,
because in this case, by chance, the order in the resulting list matches
the order in which the dictionary was originally defined.

Rewritten sentence to warn the user that one should not rely on the
order of such lists.