Return path to generated file.
This makes the gif and ffmpeg methods of an Animation return the absolute
path of the generated file.  This is consistent with the png method, which
returns the path of the generated directory.  If neither savefile nor
show_path were specified, the file name was discarded so far.  It might make
sense to depreciate show_path eventually.