Base Animation.show on Animation.save, and pass more arguments.
Now the show method is based on the save method, just like it is in the
Graphics class.  Furthermore, we pass keyword arguments down from show
through save to the final gif or ffmpeg call.  That way we won't have to
handle new arguments ar every step of that pipeline.