Turn generator into list when rendering frames.
The APNG pipeline requires the number of frames up front, and other
converters might benefit as well.  On the other hand, once the frames are
rendered (and don't get re-rendered), the number of frames is obviously
fixed as well.

As an alternative, we might save the number of frames only, discarding the
graphics objects themselves.  If someone has a good reason to do so, please
document that reason and go ahead.