Improved documentation.
The references are now HTML links. The examples take more care to get equal
bounds, so the generated frames will be of equal size.  APNG in particular
will complain if this is not the case.  The standard example of the moving
sine curve was improved to get the full range for x.  The APNG generator now
has all methods documented as well as an example up front.