AdditiveMagmas.test_zero() and Magmas.test_one() allow unhashable elements
Some elements (such as p-adics) can not be made hashable (see #11895). For
them, we can set __hash__ = None, so one could argue that the right check here
should be whether the element isinstance of Hashable instead of catching a
TypeError. However, there are objects which may contain unhashable elements
(such as a matrix over the p-adics) or not. The corresponding class then has a
hash function which raises a TypeError when including the hash value of its
entries to compute the hash value of itself.