16516: Slighlty improved the algorithm + add a algorithm="sparse" option
When the polynomial is dense, the sparse algorithm is ~1.9 times worst, it
used to be 2. For really sparse polynomials, the new version of the algorithm
is faster.

Still needs to find the right parameter to check to decide between dense and
sparse algorithm. The trade-off: Either you gain a lot (factor ~20) quite
often and loses a bit (factor ~2) the rest of the time, or you gain a lot more
rarely but almost never really loses.

Note that the tests have been made using random polynomials, not "real-life"
polynomials (that I do not know what they are!).