Create HTML to embed generated GIF in notebook.
Without this fix, a call to Animation.show() is without visible effect in
notebook, since the generated image will not get displayed.  This is because
the file is located in a temporary directory, so the notebook won't be able
to pick it up on its own.  Instead of simply dropping a file in the local
directory, the cleaner solution seems to be generating html to display it.
The approach used here is based on Graphics.show(), including the
introduction of the linkmode argument.  A common API here may help duck
typing.