Add APNG support.
The benefit here is that APNG does not add any extra dependencies.  The
downside is that only some browsers support APNG rendering.  Others would
default to showing only the first frame, but that might be confusing to
users who expect a full animation.

This is a dummy commit to share the current state of doctests.