Make number of frames available after rendering using generator.
The APNG pipeline requires the number of frames up front, and other
converters might benefit as well.  Once the frames are rendered (and don't
get re-rendered), the number of frames is obviously fixed.

As an alternative, we could turn the generator into a list when rendering,
thus avoiding the extra attribute.  But the downside of this is that the
list might be around for considerable time, eating memory.

The example animation is changed since the previous example would generate
warnings during rendering, due to negative values in the default range of x.