Fix frame size for standard example animation.
There is one animation used in many examples.  Up till now, it used to be
broken because it generated frames of different sizes.  Convert apparently
accepts this but the output looks weird, since you don't see the whole image
of some frames.  My own APNG code failed, and rightly so.