Make animation exporters return absolute file names.
Make the gif and ffmpeg methods of an Animation return the absolute path of
the generated file.  This is consistent with the png method, which returns
the path of the generated directory.  If neither savefile nor show_path were
specified, the file name was discarded so far.  It might make sense to
depreciate show_path eventually.

The claim about automatically named files getting displayed in notebook mode
is not true at the moment: the notebook will pick up images generated in the
current working directory, but the temporary files generated by default will
be within some other temporary directory instead, so notebook won't pick
those up.  So this commit removes the incorrect claim.