Slight precision gain for a p-adic L-series
The doctest does not takes less than a second on my machine, so I removed that
comment.

Here is an explanation of the precision gain.
The algorithm uses
alpha = self.alpha(prec=prec) [with prec=5]
and then computes the inverse of alpha^4
This inverse is computed using a naive xgcd which relies on a naive quo_rem
implementation.
During that quo_rem computation, we divide (in the polynomial ring) by the lift of alpha^4 which used to be
 (O(3^10))*x^4 + (O(3^6))*x^3 + (O(3^6))*x^2 +
 (2*3^2 + 2*3^3 + 2*3^4 + 2*3^5 + 2*3^6 + O(3^7))*x + (3^2 + 2*3^3 + 2*3^4 + 2*3^3 + 2*3^4 + 2*3^5 + 2*3^6 + O(3^7))
(where x is the lift of alpha.)
Now these leading inexact zero terms are gone which is why we divide by (also a
correct value of the lift of alpha^4):
 (2*3^2 + 2*3^3 + 2*3^4 + 2*3^5 + 2*3^6 + O(3^7))*x + (3^2 + 2*3^3 + 2*3^4 + 2*3^3 + 2*3^4 + 2*3^5 + 2*3^6 + O(3^7)).
This results in a slight precision gain and the result is still correct.
If you are not convinced, the xgcd in incorrect anyway, see #13439 and raises a
Stopgap Warning about this already.