Restore positional parameters.
This restores positional parameters for the sake of backward compatibility.
They might still get deprecated eventually, but with a grace period before
they get removed. Ticket #16607 might be of use there.

There also is a test case to exemplify that backwards compatibility. And I
have renamed kwargs to kwds for the sake of consistency with the remainder
of this source file.