Fixed package checksums munching by /src/bin/sage-fix-pkg-checksums
Also spotted a couple of possible enhancements, which are for now just
presented as suggestions to users : it should be possible to carry the
necessary modifications automagically. This should be evaluated by
someone that knows his|her way in the sage development system (I
don't, really) and knows what {s}he is doing (I don't either).