Trac #13569: Revert basis of quaternion algebras and orders to be tuples
Using lists for cached attributes like basis is either dangerous (if you
return without copy) or slow (if you return with copy). Storing the
basis as a tuple ensures that it is immutable and can thus be returned
by reference (i.e. without copy).

This also makes the total patch way simpler.