preparse_file: trigger SyntaxError on assignment to numeric literal.
The core idea is to replace numeric literals not directly with the
corresponding variable name, but instead with a formatting placeholder
similar to those used for string literals and comments.  That way we can
substitute different things for different applications: for syntax checks we
substitute real python numeric literals, namely the literal "1".  For
compilation and execution, we plug in the variable names.  This approach was
suggested by Robert Bradshaw.

This change deliberately breaks compatibility with regard to the values
returned by preparse_numeric_literals(code, extract=True) and
extract_numeric_literals.  Where the first element of the returned tuple
used to contain variable names, it now contains formatting termplates.
There are also some slight deviations regarding spaces added after raw
Python numeric literals.  It is assumed that no user code will rely on any
of this functionality, except indirectly via preparse_file.

This change might incur some performance penalty since the Python code has
to be parsed twice.  On the other hand, there are some situations where
performance should become better.  We avoid extracting string literals
twice, in preparse_file and preparse.  We also avoid running the preparser
on the numeric literals.  The latter apparently was a limiting factor for
the number of assignments which could be placed in the first row.  Perhaps
that limit no longer applies now, and should be removed in a subsequent
commit.