Increase precision when converting to radical expression.
Instead of casting the resulting symbolic expressions to algebraic numbers
and doing an exact computation on these, we double the precision of the
interval field until only one of the found symbolic loops intersects our own
value interval.  This avoids the need for a separate code path, and the
existing doctest about the double overlap should cover this situation as
well.

As a side effect, we no longer have a code path to deal with the case where
the list of symbolic roots contains less roots than the degree of the
polynomial.  At the moment there is no known case where this would happen
(short of an empty list if no radical expression could be found at all).