Avoid code duplication between qqbar and nfe.
The conversion of number field element generator embeddings to symbolic
expressions is now delegated to the radical expression conversion of
algebraic numbers.  If that conversion fails, e.g.  because the degree of
the minimal polynomial is too high, then we'll now end up with an algebraic
number embedded into the symbolic ring, where before we'd have raised an
exception.