Convert qqbar to radicals via number field.
As discovered in the example which is now included in the doctest,
converting an algebraic number to a number field element can lead to a
situation where the defining polynomial of said number field can be solved
symbolically even though the minimal polynomial of the algebraic number
could not.  This gives us a second chance at finding symbolic expressions
for algebraic numbers, so we try that if we got no (or not enough) solutions
for the minpoly.  Since the resulting expressions can be pretty large and
ugly, we introduce a flag to control this step of the conversion.

When converting number field elements to symbolic, we prefer not to
introduce a new number field, but operate in the input number field instead.
So the first call to radical_expression has via_nf=False.  When converting
the generator, though, we want to try very hard to find a conversion.  So in
that case we let it fall back to a new number field if that helps.