Build symbolic expression from descriptor graph of algebraic number.
The current state of affairs is probably going too far: we manage to
construct symbolic expressions for many algebraic numbers, even if there was
absolutely no symbolic expression involved in the input.  In many cases,
the resulting exactification is different from what had been used so far.
For this reason, many doctests in qqbar (and likely other places as well)
currently FAIL.  Some of the discrepancies are for the better, though.
We have yet to decide where to draw the line, when to go via symbolic
expressions and when not to.