Fix hidden bug in RDF/CDF matrix multiplication dead error-checking.
AX = B implies A.nrows() = b.degree(), not A.ncols() = b.degree().

[n x m][m x p] = [n x p]

Since the matrix multiplication for RDF/CDF matrices only works when A
is square, and thus when A.nrows() = A.ncols(), this bug was hidden.

(Note that b is a vector and A is self.)

Also, this error check is dead code. If b is not of the correct degree,
it will not be successfully coerced into the column space of A, and the
solve_right/solve_left code will error out before this error check is
performed.

Maybe it should be deleted?