Fix trac #18032 by making sandpiles copyable.
The generic graph class GenericGraph defines a __copy__ method which is
intended to work for any of its child-classes. Unfortunately, the
sandpile class Sandpile (which inherits from GenericGraph through
DiGraph) changes the signature of the __init__ method and breaks
__copy__.

To fix this, I have moved the part of __copy__ that depends on
__init__'s signature into its own function in GenericGraph, and
overriden it in Sandpile.

In addition, Sandpile's __init__ was not correctly piping extra
arguments to its parent class' __init__ method. I have fixed this as
well.

Add tests to Verify that sandpiles can be copied, and that they cannot