Unify key domain for PolynomialSequence_gf2.solve
Now the key domain is the set of monomials from the appropriate ring.

The key conversion function, which is given as a lambda, first converts to
the polynomial ring, then takes the leading monomial of the result.  This
might cause strange results if the provided key is not a bare monomial.

Some doctests were made less random by using reproducible_repr.