Trac #18454: Allow random_cone() to be (non-)strictly-convex.
This commit adds a parameter, "strictly_convex", to the random_cone()
function. It takes three values, determining whether or not the
returned cone will be strictly convex or not. The third value, None,
means "I don't care."

Two additional sanity checks are introduced that prevent infinite
loops (when the user asks for something impossible). Many more
doctests are added to test both the desired behavior and the error
conditions.

Finally, the random cone generation algorithm has been updated. We
generate cones in a loop as before; but now, we try to "fix" the
generated cone to make it (non-)strictly convex. If that doesn't work,
we throw the cone away and loop per usual.