Trac #18454: Add a "solid" parameter to random_cone().
With the addition of the "solid" parameter, the random_cone() function
can now generate (non-)solid cones at the user's request. The
algorithm for cone generation did not need to change at all: we rely
on the fact that if we wind up with a dud, we simply throw it away and
start over. We do however massage max_dim/max_rays if the user has them
set to "I don't care."

Many more test cases are added, as well as warnings about the slow
generation speed when the user becomes too picky. The majority of the
code for the feature comes from the sanity checks associated with the
parameter.