Changing the ducktyping code
I don't like the idea of checking that loc is an integer by calling self[loc]. Maybe one day we'll have a shortcut self[i, j] for the (i, j)-th entry of self, and then this will break. I prefer to check whether loc is a pair by means of "r, c = loc". The only way I see how this could fail is if an integer is given as an element of a combinatorial free module, as in "s(3)" where s is the Schur basis of Sym. But this is bad user input IMHO.