Trac #19044: Add span() method to IntegralRayCollection.
The existing top-level span() function works on cones, but will fail
on a trivial cone since it is unable to deduce the underlying
ring. This commit adds a new span() method for IntegralRayCollection
(and thus for cones) that addresses this shortcoming. The new method
uses the span() method of the underlying lattice -- this gives
us a sublattice in the common case, but still allows us to obtain a
vector space by specifying a base_ring.