Trac #19332: Add discrete_complementarity_set() method for cones.
Any sort of "linear property" on a cone can usually be verified on a
generating set of the cone instead of the whole thing. This involves a
finite number of steps for a polyhedral cone.

The discrete complementarity set is a finite subset, consisting of
generators, of the usual complementarity set that appears in
complementarity and mathematical programming problems. Having it
available lets us check complementarity properties on the cone.

This commit adds the method, examples, and tests.