Trac #19332: Use reviewer's implementation of discrete_complementarity_set().
The initial implementation of discrete_complementarity_set() converted
all rays to vectors before computing inner products and returning
orthogonal pairs. Now the inner product is computed directly on the
rays of the cone. As a result, toric lattice elements (not vectors)
are returned.

This simplifies the implementation, and the test output has been
adjusted to reflect the ambient lattice of each ray.