Trac #19485: make legends more legible.
The default "legend options" are empty, but this defaults everywhere
to a gray legend with black text. This is difficult to read,
especially when the result is printed (journal, thesis, etc.) or
projected (as in a presentation).

This commit changes the default background color for legends to
'white'. A drop shadow is added to ensure that the legend stands out
amongst the plot.

Some doctests and examples have been updated to make sense considering
the new defaults.