Trac #19405: add strict_quotient() tests and use it in lyapunov_rank().
One of the two calls to _restrict_to_subspace() in lyapunov_rank() can
be replaced by a call to the existing strict_quotient() method. Some
of the examples and tests for _restrict_to_subspace() also apply to
strict_quotient(), so they have been copied there. After moving the
tests, the call itself was replaced.