Trac #19831: add "ring" parameter to random_element() for cones.
At the suggestion of the reviewer, the random_element() method for
cones now returns a lattice element by default. The method however
takes a "ring" parameter, which when set to "QQ", causes the returned
element to live in the ambient vector space instead of the cone's
lattice. The tests and examples are also updated.