Trac #19831: change ValueError to NotImplementedError in random_element().
In both toric geometry and optimization, cones live in real vector
spaces. It is an implementation detail that the ambient vector space
in Sage is rational. The existing `ValueError` is therefore changed to
`NotImplementedError` when the coefficient ring of `random_element()`
is not rational.

Some minor documentation updates are made as well.