Altered num_faces code
Reverted some of the white space changes that I made in an earlier patch, the raw_string marker was making all of the code below it appear to be commented for some reason.
Removed the ValueError catch as it is automatically raised by calling faces on a graph without a planar embedding.