Replace _cache_key with strict_equality
I tried to make strict_equality as fast as possible. I need to use a
stack instead of just a python list for it to be fast.
Here are some timings. In all these examples, `X =
toric_varieties.dP8()` and we consider `%timeit [X.cohomology_basis(1)
for i in range(1000)]`.

The current `cachefunc.pyx` (using `_cache_key` in a try-finally-block):
1000 loops, best of 3: 419 µs per loop
The old version which does not handle unhashable objects at all:
1000 loops, best of 3: 387 µs per loop
The proposed implementation:
1000 loops, best of 3: 395 µs per loop