Enable unpickling of legacy p-adic elements
p-adic elements used to be hashable. They are not hashable anymore. This
is a problem when they are part of a dictionary in a pickle. We have no
way of intercepting the dictionary creating in a pickle so they need to
remain hashable. Therefore we keep track (globally) of all the legacy
p-adics that come from old pickles.