Improve reporting of Features
Introduced a FeatureTestResult class which can be used as a boolean.
This seems to be the most elegant way to keep is_present() functional
(i.e., returning something that behaves like a bool) and attaching
diagnostics to it.
An alternative would have been keyword arguments such as
is_present(explanation=True, resolution=False) and returning a tuple if
any are present but I think that this is rather annoying for the caller.