Handle setting permissions of DOT_SAGE better on Windows.
For one, this code should only be executed if os.chmod exists,
which it does not normally on Windows.  Granted, we are not
currently targeting native Windows support, but for the future
it's nice to check.

At any rate, this often fails quietly on Cygwin as well.  This
is because Windows filesystems can be mounted in Cygwin with
or without ACL (Access Control List) support--Windows' (and hence
NTFS's native filesystem permission system).  With ACL support
enabled, Cygwin's FS layer tries (and does a reasonably good job)
at emulating UNIX-style permissions through the ACL.

However, if a drive is mounted with no ACL support, most attempts
to chmod a file or directory will either result in a no-op, or
at least not do exactly what was expected (minimal support for the
read-only flag is provided, but not much else).  Therefore, setting
the requested permissions on DOT_SAGE will quietly fail, resulting
in repetitious "Setting permissions of DOT_SAGE directory" messages
every single time sage.misc is imported.  This also causes several
tests to break.

As a workaround, when the permissions are *not* actually being set
we just remain quiet about it and don't tell the user that something
is happening when it isn't.