Adds patch to pip to allow it to be used without SSL support from Python
This applies a patch [1] to pip's vendored copy of distlib, which was
previously responsible for breaking pip when Python's ssl module is
not available.  The rest of pip appears to work without SSL support,
at least when performing operations that don't require an HTTPS request.

A new release of distlib is needed before pip will upgrade their
vendored copy.  But then we can remove this patch.

[1] https://bitbucket.org/vinay.sajip/distlib/commits/a258e3bdd6f8