Fix zip file opening.
ZipFile.open has very different meaning from TarFile.open.  ZipFile.__init__ and TarFile.__init__ are fairly similar, but the latter does not have the higher-level features of TarFile.open, namely transparent handling of compression.

This is a bit hacky, but it allows SageTarFile.__init__ to take full advantage of the TarFile.open classmethod.