Add a new -y argument to sage-spkg to make it automatically answer 'y' to all prompts.
Changes build/make/deps so that when running make, 'sage-spkg -y' is the default.  On the other hand 'sage -i' does not pass the -y argument to sage-spkg, so it still prompts the user for action.

This also totally redoes the prompts to be more sophisticated than just using 'read' directly, which isn't very good.  The biggest advantage to this is that the user's answer to the prompt is logged.  This is also the only way to get the newline entered by the user to be picked up by the logger (otherwise it will fail to add a prefix to the next line output after the prompt).