Warn but don't fail if there are no patches to apply
I think the wording of the warning is such that it shoudn't make anyone panic--if there are no patches to apply that's fine, that's normal.  But I think it's also helpful to have especially if you're developing an spkg and *expect* patches to be applied--this will tell you you have something wrong.