Reworked jmol installation to be somewhat more consistent.
There was one file (JSmol.js) that was previously being replaced, but not any longer, so it can be deleted.  Another file, the jmol executable, was not previously patched by appling a patch to the upstream sources; rather, we included an entire copy of the file which we copied over.  I don't see any particular reason for this so I changed it to simply apply a patch (and got rid of the patches that are no longer used).