Make tachyon conform to the standard patch format.
There was even a "TODO" item in the commit message
that first added the patches for this package to apply
them as patches rather than including full copies of
the files.

The patched versions are now quite diverged from the
upstream versions of those files, so the patch files
themselves are now different from the original versions.
But I kept the patched versions of the files (i.e. the
end result) the same, since it seemed to be working anyways.