Adds two new patches for ECL to work properly on cygwin.
These patches are mostly unrelated.  The first is an issue where the
`uname` built-in function is not defined in cygwin, even though it
should be available. This leads to a segfault in code that tries to
use it.

The latter is a linking error on Windows caused by a simple omission,
and is already fixed upstream (but we still need the patch for now).

Neither of these issues cause ECL to fail to build/install, but do
lead to installation failure of Maxima.