Fix scipy/sparse/sparsetools/bsr.cxx in debug builds on Cygwin
This module does not build on Cygwin with the default options due to size added to the binary by the debug symbols.  Adding -Wa,-mbig-obj allows the assembler to use extensions to PE32+ to use larger sections in the binary.

Also the logic of this change is such that the -fPIC flag is removed on Cygwin which is desirable (see
https://trac.sagemath.org/ticket/20508)