Allow SciPy to know that Numpy was built with ATLAS support.
Setting ATLAS=None forces Numpy (and by extension SciPy) to think that ATLAS is not being used, even if it is the BLAS library we're using (as is default for Sage).  In this limited case, where our blas is linked using -latlas, we can be pretty sure ATLAS!=None.  See https://trac.sagemath.org/ticket/20507