NTL: Do not entirely disable NATIVE when Sage's GCC is used, white-list some assemblers instead (including LLVM's).
If the *default* assembler on Darwin is still Apple's ancient
version of the GNU assembler, we do not (yet) try to use the
alternate one from LLVM (which could be achieved by adding '-Wa,-q'
to CFLAGS and CXXFLAGS); such should probably be done "globally"
for Sage, e.g. by putting a wrapper script into $SAGE_LOCAL/bin/
when appropriate.  Something left for #20779...