Use this TemporaryFile as a context manager so that it's closed/removed as soon as it's done with (without waiting for garbage collection).
For some reason leaving this particular TemporaryFile instance open was causing test failures, because the file would be deleted before it was closed, leading to an unhandled OSError in its __del__ (at least on Windows).

I was never able to work out why the file was being deleted, as nothing directly in the Sage code manually removed the file.  It's also possible the problem could exist in other parts of this code, but I wasn't able to reproduce it after making this one change.