Work around some non-deterministic failures in these tests by synchronizing the expect interface first.
Not 100% sure if this solves the problem, as the problem only occurred rarely in the first place (sometimes the first _sendstr would not complete before the second command).  Forcing a sync seems to make a difference as I have not been able to reproduce the problem since, but can't be certain either.