Make this test less sensitive to minor numerical differences.
Small platform-specific differences in floating point rounding for the first line in this test could
lead to subtly different polynomial results for algdep.  At worst there would be an additional factor
(in this case of x^2).  So don't test the representation of the polynomial itself; just test that it
contains some expected factors, and that the input to algdep is still a reasonably approximate root.