Add a sage-pip-install wrapper for pip which handles uninstallation better.
pip install normally uninstalls any conflicting packages (i.e. older versions of the package being installed), but it does not do this with --ignore-installed (which is the easiest way to force pip to install a package that is already installed).  This is annoying, so we first manually uninstall the pakage (multiple times, if necessary--one feature of .egg files is that it's possible for many of them to be hanging around in the user's site-packages).

This resolves #21441