Trac #21595: add cross_positive_operator_gens() method for cones.
This commit actually adds two new methods. The first new method is
cross_positive_operator_gens(), which computes generators of the
cross-positive operators on the cone whose method it is. The second
method is Z_operator_gens(), and is unremarkable because the
Z-operators are simply negations of the cross-positive operators.