fix charpoly(algorithm='flint')
charpoly(algorithm='flint') was messing with the generator of the
polynomial ring. This is fixed. Also, this commit adds an actual
call with algorithm='flint' in a test, and removes an old docstring
that claims that linbox isn't used for charpoly() on 64-bit systems,
since it certainly is used on my 64-bit systems and I can't find
anything that would set it to not be used.