Trac #21595: remove unbounded random integers from doctests.
The embarrassing cause of the intermittent test failures (due to
timeout or resource exhaustion) turned out to be an unbounded call to
ZZ.random_element() that was copy and pasted into a few doctests. The
result was that a trivial cone could be created in a huge lattice,
starving the machine of CPU and/or memory.

The fix restricts those calls to integers between zero and four. In
simple cases (like the trivial cone), sixteen dimensions is not a
problem.