Trac #21595: add a citation for some (cross-)positive operator tests.
Many of the doctests for the new positive_operator_gens() and
cross_positive_operator_gens() methods test properties like their
lineality and dimension. These are not published anywhere yet in
the general closed and convex case, so I've provided a preprint
reference for them:

  [OrlitzkyPosZ] M. Orlitzky.
  Positive and Z-operators on closed convex cones.
  http://www.optimization-online.org/DB_HTML/2016/09/5650.html

The algorithms themselves appear in the same paper, but are trivially
true in the polyhedral case where everything is closed.