Trac #21595: remove redundant random positive_operator_gens() tests.
The doctests for positive_operators_gens() contained a few cases that
are redundant. We do wish to test all of the following properties,
which had doctests:

  * Each positive operator should send the generators of one cone
    into the other cone.

  * Each positive operator should send a random element of one cone
    into the other cone.

  * A random positive operator should send the generators of one
    cone into the other cone.

However, given enough runs, it suffices to check that a random
positive operator sends a random element of one cone into the
other. Such a test already exists, so the three tests mentioned
above have been removed.