Fix derivative of gen_laguerre wrt non-implemented variables
The problem was that the return value was in the 'else'
branch of diff_param, so anything different than diff_param == 0
would return the derivative wrt x.

The proper way to do this is to actually check for diff_param == 2
to return the derivative wrt x. When diff_param == 0 or 1 we
print a "non-implemented" error message.

Signed-off-by: Carlos R. Mafra <crmafra@gmail.com>