Normalize arguments of AlgebraMorphism
and do not abuse asserts.

Note that this slightly changes the meaning of ==. Now, two morphisms which are
in different categories (but whose __dict__ is the same) are different. Before
they were ==. I think that this makes sense. x == y should imho be True iff the
two items are indistinguishable. If they live in different categories, they are
certainly not indistinguishable.