Updated terminate() to add a little more customizability in its behavior, with I think reasonable defaults.
I also updated the tests to only print() once. The reason for this was that, at least on the
python I'm testing on currently, killing the process results in a broken pipe exception *from*
the subprocess, which also breaks the test.

It still prints once as a means of ensuring that the process is fully initialized before trying
to terminate it.  Then it just goes into a dead loop.