{c,sparse}_graph: systematically turn integer-like vertices into ints
Instead of doing it or not depending on the phase of the moon:

sage: g = DiGraph([[1..12],lambda i,j: i!=j and i.divides(j)])
sage: [type(v) for v in g.vertices()]
[<type 'int'>,
 <type 'int'>,
 <type 'int'>,
 <type 'int'>,
 <type 'int'>,
 <type 'int'>,
 <type 'int'>,
 <type 'int'>,
 <type 'int'>,
 <type 'sage.rings.integer.Integer'>,
 <type 'int'>,
 <type 'int'>]

Leaving the vertices alone would be a better fix in principle.
Unfortunately, it would break compatibility too badly: for example,
people clearly expect to be able to create graphs in library code, using
Python ints as labels, and then access the vertices using Sage integers.