Adds a sage_build_py command in setup.py
This uses the reworked find_extra_files, along with a (currently hard-coded)
data structure more or less identical to the distutils package_data argument,
to copy supplemental non-Python files into the installed sage
package/sub-packages.  This list is also passed to to cleanup_stale_files as
with the extra files found during the Cythonization step.

Most of this could be avoided if we just used package_data directly but
supposedly that's broken in some case.