Initial refactoring of sage_setup.docbuild.__init__ into a number of submodules.
The specifics of what went where are mostly subjective and done by feel, with
one exception: I wanted to make sure it's possible to at least import some of
the builder base classes without having to import all of Sage.  For example I
wanted to be able to run AllBuilder.get_all_documents without all of Sage being
importable yet.  This is for the purposes of another issue (#21732) and some
other issues related to it.

Aside from some minor stylistic changes (imports, whitespace) there are a
handful of small substantive changes in this commit:

 * AllBuilder.get_all_documents takes a default_lang option.  This implements
   the behavior previously hard-coded in the get_documents() helper function,
   that stripped the 'en/' off the names of all English language documents.
   This is mostly with a mind toward some future changes.

 * The 'options' global variable is done away with, and all global options
   are accessed from the build_options module.  This is just to improve
   consistency in how build options are handled.  The use of a bunch of global
   variables in an options module has some code smell in its own right, but I
   otherwise left that as is for now.