Don't mix bytes and strings sloppily.
Do use universal_newlines=True with subprocess.Popen; this is usually
the easiest cross-Python way to wrap I/O streams on a subprocess as text
streams, using the preferred locale for the system.  There are cases
where this can fail, such as when the system's locale is misconfigured,
but this generally not a problem for plain-ASCII interfaces.