Add some link flags needs to build some modules on Cygwin
On *nix it is possible to link shared libs without resolving all external symbols at link time;
however Windows is more strict on this and requires all symbols to be resolved at link time,
hence requiring one to more explicitly list the libraries a module depends on.

In any case it doesn't hurt to be more explicit.