Trac #22415: add operator class membership tests for matrices.
This commit adds four new methods for matrices:

  * is_positive_operator_on
  * is_cross_positive_on
  * is_Z_operator_on
  * is_lyapunov_like_on

Each new matrix method tests whether the matrix in question has a
certain property on a polyhedral closed convex cone. The properties
are variously described:

  * Positive operators send one set into (a subset of) another set.
  * Z-operators generalize Z-matrices by allowing nonstandard cones.
  * Cross-positive operators are simply negations of Z-operators.
  * Lyapunov-like operators are the intersection of the Z-operators
    and the cross-positive operators for a given cone. They generalize
    the Lyapunov transformations one sees in dynamical systems.

The new methods are implemented naively, by checking some
inequalities. This allows the user to use matrices with symbolic
entries, since (for example) checking whether or not a symbolic matrix
is Lyapunov-like might reduce to checking whether or not ``pi - e == 0``;
namely something we can actually decide. If speed is needed and the user
has rational matrices, he can check them against (for example) the set of
``positive_operators_gens`` for his cone.