Trac #22415: use new matrix methods in cone tests.
Three cone method tests for lyapunov_like_basis(), Z_operators_gens(),
and cross_positive_operators_gens() were testing that some operators
are Lyapunov-like, Z-operators on, or cross-positive on a cone. Now
that we have the e.g. is_lyapunov_like_on() method for matrices, it
makes sense to replace the existing cone tests with a call to it.

Once that is done, a few matrix tests are partially redundant. There's
no need to test the same matrices that the cone methods already test,
so those doctests have been removed. The tests involving the symbolic
ring, however, are still relevant and have been left unmolested.