Add patch to psutil for supporting Cygwin
The patch is currently something of a mess.  In particular, it duplicates
a lot of code from the Linux and Windows support modules.  I plan to
continue improving it but by the time that's done it will probably be merged
upstream shortly after, at which point a patch will no longer be needed.

In the short term I need this resolved as it is a major blocker to restoring
Cygwin builds.