Fixes a bug with reading from stdin on Cygwin (and possibly other platforms)
When fflush() is passed NULL (i.e. if pari_logfile is NULL) it flushes *all*
open streams.  For input streams this means clearing any unprocessed input
from any buffers.  This can cause fgets to return an early EOF.  This isn't
a problem (it seems) on Linux, which provides additional behavior in fflush
to not discard buffered data from non-seekable files (such as piped in
stdin), but this is not POSIX standard and can't be relied on.

fflush(NULL) is not what was intended here anyways so just check that
pari_logfile isn't NULL first.