Using AppleScript to set background image of the .app.dmg (for Mac installations) instead of the ds_store python package by Alastair Houghton.
Mac OS 10.12 (Sierra) changed the proprietary .DS_Store file format (which contains the information where the background image is and how to position the icons) so the ds_store package broke.

The advantage of AppleScript is that we are isolated from the details of the .DS_Store file and that the AppleScript API is hopefully more stable and reliable than trying to create the .DS_Store file itself which is proprietary and undocumented.
The disadvantage is that this requires that we first produce a read-writable disk image and can compress it to a readable image only after we have mounted the disk image to run AppleScript to set the background image. This means that the build time is slightly longer and that we create a intermediate file several Gb in size.