Modify the Dokchitser class to only require one gp interpreter which can be shared by all instances of the class.
This is not straightforward to do because Dokchitser's script
(compute1.gp) uses a number of global variables, and functions that work
on those global variables (understandably so).  So it would not be
possible for multiple instances to exist simultaneously without having
the reinitialize the computation for each evaluation of the L-function.

This is achieved by amendng compute1.gp with a template file
compute1.gp.template, written in the syntax understood by Python's
built-in string.Template.  This was chosen in part for its
simplicity/availability, and the fact that its syntax uses dollar-signs
to identify template variables, which are not otherwise used anywhere in
the script (thus obviating any awkward character escaping).

The only template variable in the template is $i.  Every global
variable and/or function that uses any global variables is replaced with
<varname>_$i (the underscore is required for clarity and disambiguation,
as some function names end in numbers).  $i is always replaced with an
integer, monotonically increasing for each instance of Dokchitser.  So
for example in order to initialize the script one calls initLdata_<N>
instead of initLdata.  And in order to compute a value of the L-function
one calls L_<N>(s) instead of L(s), where <N> is the instance number.

This is handled transparently by the class implementation so of course
there are no changes to the user API.

This provides a work-around to https://trac.sagemath.org/ticket/22718