Updates the autotools spkg to use staged install
This was actually rather tricky to get right, since the autotools spkg actually installs multiple
packages, each of which depends on some of the previous ones at build time.  The only real way around
this when installing to a temporary root is to make the temp root the actual install root of the
packages (and include it in PATH).  Once everything is built we then do a mass sed to replace
all references to the temp install location with just SAGE_LOCAL, which seems to work.

The only remaining references to the temp install root after running sed are in a handful of binaries,
and only in the debug info at that (they still have the correct rpath, for example).