Remove this sanity check
We are safe in assuming that patch's make install works correctly, and no
explanation for this check was given in the commit log when it was first
added (it might be because the no-op spkg-check was removed, but it's still
unnecessary).

If really needed this could be added back in as a post-install script once
that functionality is added.