Remove _cache_key
Arguably, str(self) is not a good default to identify objects. It often
does the right thing, but it also causes trouble when objects print the
same, are == but are not indistinguishable (and should therefore not be
identified in caches.) It seems to be better to require people to
actually implement _cache_key explicitly, just like they have to
implement __hash__ explicitly now.