Create wrapper script around spkg-build, spkg-check, and spkg-install
This prevents scripts that already contain a shebang line from running at all, just as a sanity
check (i.e. they should be written with the assumption that they will be wrapped later, and should
not be executable on their own).

So with this change no packages will install until they've been updated.