As suggested, rename sdh_check_common_vars to sdh_guard
Add it to the standard script header. Not 100% sure if it would be better there, or
explicitly in every script file.  In the automatic header it's mostly redundant--if
sage-env succeeded then it's not very likely needed, unless sage-env itself is broken.

So it probably doesn't hurt to check as a last defense.