Add support for building with 'abstract' packages.
Add 'mp', 'blas', and 'python' as abstract packages, replacing the old
Makefile variables 'MP_LIBRARY', 'BLAS', and 'PYTHON' respectively.
Packages that used those variables in their dependencies now depend on
the corresponding abstract packages.

In terms of functionality this is not much different from the old
implementation.  Mostly it just does a better job of keeping track of
what selections were made for a given abstract package, and opens the
door to removing more special cases and other new functionality based
on this.