Use AC_CONFIG_FILES + AC_OUTPUT to generate build/make/Makefile
This replaces all the logic to generate build/make/Makefile directly in
the configure script, with essentially the same logic, but instead it
generates variables that will be substituted with AC_SUBST into a
build/make/Makefile.in template.

That template in turn uses eval loops to generate the rules for individual
packages, rather than explicitly writing all the rules in the Makefile.