Further rewrite of configure.ac to do away with the multiple filtered_packages_list loops.
With the previous simplifications it makes more sense to just loop
over the package list once and update the relevant variables.  This
simplifies the code somewhat, especially doing away with the awkward
piping from heredocs, that was necessary in the previous commit in
order to avoid piping to a subshell.

This is actually *noticeably* faster, especially on Cygwin where running
configure was always very slow due to the large number of subprocesses
involved.

This also corrects a bug from the previous commit with handling standard
packages that are not installed based on the configure result (e.g. gcc,
git).  It should still be possible to install such packages manually.