Convert several autotools packages to using the helper functions for configure/make/make install.
These are all "well-behaved" packages in the sense that they are all have
straightforward build scripts that enabled a completely mechanistical
replacement of "configure" with "sdh_configure" and so on, as well as deletion
of no-longer-necessary error-handling boilerplate.  This also removes the
"SAGE_LOCAL" guard from the top of each of these scripts, as it is no longer
needed since #23179.