Add the ability to temporarily install to an alternate installation root (referred to often as the DESTDIR convention, which is supported by most automake packages).  From the temporary install root files are then copied into the real install root.  This is called a "staged installation".
This has a number of uses, such as being able to build a binary archive
containing just the files installed for that package.  Here we use it also
to generate a manifest of files installed/owned by a given package.  This
can be used, for example, to implement more careful package uninstallation.

In particular, this supports passing a SAGE_DESTDIR variable to spkg build
scripts.  In the case of packages installed with `make install`, this is
passed to `make` as `DESTDIR=$SAGE_DESTDIR`.  Not all packages that use
make support this (but most do).  For those that don't this is a no-op.

The same could be done for Python packages by having `sdh_pip_install` add
`--root=${SAGE_DESTDIR}`, but currently that breaks a tiny set of Python
packages whose build scripts do not fully support this kind of staged
installation, so their case will be addressed separately.