Better support on Python 3 for doctests that expect exceptions.
As the existing tests were written for Python 2 most of them fail on Python 3
since in the latter exceptions are usually printed with their fully qualified
name (module name plus class name).  Here we allow a little flexibility so that
the tests can still pass on Python 3, as this repr difference does not impact
the validity of the test either way.

Conflicts:
	src/sage/doctest/forker.py