Workaround a noisy doctest
we had seen issues with this test at some point
(https://trac.sagemath.org/ticket/21869#comment:274.) We are not sure what's
the issue and probably it is not even caused by something random happening in
the valuations code. Anyway, we relax the check here to make sure that others
don't see rarely failing doctests (in the patchbots) for an area they do not at
all care about.