As an initial demonstration of the new system, move configuration for yasm to an spkg-configure.m4.
For the case of YASM we only require it at all for certain architectures
(albeit x86(-64) so this is most of the time.  So this first checks
whether we require YASM at all.  It then performs the necessary feature
check as to whether we can use the system's YASM (if one even exists).