Slightly rewrote this macro to use some more polymorphic shell macros.
In particular, also only sets sage_spkg_install_foo and sage_require_foo
if they were not already set.  This is necessary because some macros
required by individual spkg-configure.m4 checks may be written to configure
out of the expected order (this was the case, for some reason, with
AC_FC_FREEFORM, for example).