Improved (?) documentation of the Makefile.in template.
Made the templates themselves look a little less like line noise by passing in
more pre-expanded variables (namely, the package version, and the list of
package dependencies.  This makes the template itself look a little cleaner at
the cost of slightly more complexity when calling the template.

Moved dependency expansion into its own macro to reduce a bit of duplication,
and also fix handling of order-only prerequisites (previously the the | symbol
indicating order-only prereqs was being expanded to $(inst_|), oops.

Added the DEBUG_RULES flag, so that running the Makefile with this flag prints
all the generated rules for inspection purposes.