py3: some fixes related to properly handling division, particularly in sage_setup.autogen
Previously we used an opcode named 'div' both for the old div operator
and for truediv (for compatibility's sake, since use of __truediv__) depends
on whether or not `from __future__ import division` is in effect).

So again, for compatibility's sake, on Python 3 we call truediv just 'div',
while integer division is still explicitly 'floordiv'.