Fix doctests
I thought for a while that the behaviour of these doctests was random. Actually
it is not. It is consistent over all patchbot runs with the same base version
of Sage. Caching the polynomial rings flips some signs, I am not sure why.
Anyway, the result is still correct and it is not random, so there is no need
for a `_test_S_units()` which would be hard to write anyway.