Adds a new sdh_install helper function:
This is like a very simplified version of the `install` program, and is
intended for use by packages whose spkg-install just copies some files (or
more complex packages that copy some files as one of its steps).  The
majority of its functionality is just in ensuring that the full destination
path exists, and to provide some error checking.  This replaces some
patterns that occur frequently in some spkg-installs.

Also updates the mathjax and thebe packages as two examples of its usage.