Drop tachyon's usage()
I don't think it's necessary to expose this in the Sage interface. This fails
for Debian at least (where tachyon's usage screen has been patched?) If people
are in IPython or a Jupyter notebook, they can just type `!tachyon` to see the
same information. In theory this breaks the existing interface but I am quite
sure nobody is relying on this.