Remove all runtime dependency on a VERSION.txt file
Instead the string that is written to VERSION.txt also lives in new variables
added to src/bin/sage-version.sh and src/sage/version.py.  It is slightly
redundant but this is tolerable.  The template for this string now lives in
sage-update-version rather than in the Sage library itself, which I think
makes sense.

The VERSION.txt file itself is still kept since there are some external
utilities (e.g. the patchbot) that read it.