Trac #25199: fraction field hash code no longer assumes hash(1)==1
The goal is to make sure that fractions with 1 in the denominator
produce the same hash value as elements from the underlying ring.

Instead of testing if hash(d)==1 (d is the denominator), we now
test if d==1.  This will slow down the hash because it actually
tests for equality, instead of merely testing for hash equality.