Run (--long) doctests as part of CI
on CircleCI we run the non-long doctests always the about half an hour, if I
recall correctly. On GitLab the non-long doctests take about 90 minutes on the
shared runners which is too much to run them always. The --long doctests takes
ages on the shared runners, so we only run them when they were triggered
manually.