Change the repr formatting to ensure that if the keys argment was given, the elements given in that list are sorted accordingly (and any remaining elements are appended to the end).
Unfortunately pformat does not provide a means of specifying a sort key
directly.  We could work around this with some kind of dict wrapper that
specifies the ordering for items() or using an OrderedDict.  But unless
this case comes up more often, for now it's simple enough to roll our own
solution.