Refactoring of classes in sage.algebras.quatalg.quaternion_algebra along the same lines as https://trac.sagemath.org/ticket/25387
However, this adds UniqueRepresentation to some classes that didn't already
use it: `QuaternionAlgebra_ab` and `QuaternionOrder` both of which I think
can be benefit from UniqueRepresentation.

It also adds UniqueRepresentation to QuaternionFractionalIdeal_rational
thanks to https://trac.sagemath.org/ticket/25388, which again seemed
reasonable (in this case these are only unique for a given basis which
is consistent with how these elements are compared).