Necessary follow-up:
This was sorting instances of QuaternionFractionalIdeal_rational which don't
actually have a useful ordering defined (they were just being sorted by id()).
This happened to give reasonably consistent results before, but now that they
are cached as UniqueRepresentations that was no longer true (and it was just
by luck anyways).  Instead we explicitly use their bases as the sort keys.