fix coercion bug and override methods on X, Phi and Q bases
The coercion bug fixed here predates this ticket; it is due to
the transience of the bases of WQSym. I have now made the bases
permanent (as long as WQSym lives).

The algebraic_complement, coalgebraic_complement and
star_involution methods have now been overridden on the X, Phi
and Q bases of WQSym.