Trac #25351: enhanced @handle_AA_and_QQbar to handle factorization
The decorator @handle_AA_and_QQbar now takes an optional argument
factor_field indicating that a number field large enough to factor
the arguments should be used.

The decorator has been added to various polynomial ideal methods
that can now be supported with this enhancement.

This enhancement does depend on Trac #25390 in the sense that it
won't work without the ability to factor multivariate polynomials
over QQbar, but an exception will be thrown if the factorization
fails, so no incorrect results are given and the behavior is little
different than it was before (i.e, the methods didn't work and
threw exceptions).