Add tests for 94cd188
these valuations do not really support comparison yet, i.e., there is nothing
smart yet to find out that [v(x)=1] after x→x/p is the same as the one
induced by the Gauss valuation. It is not clear whether that would really be
important though so we'll try without for the time being. (These valuations
are certainly not what most people have in mind when they think about
valuations on rational function fields anyway.)