slight changes to (still kludgy) command line handling
Removed explicit -gthread -qthread -q4thread -wthread options as the are no
longer supported by IPython.  Also removes explicit documentation of -pylab
option.

However, this preserves a generalization of this semantic, in that if either
no argments are given, or the first argument is an optional flag, we launch
the interactive Sage session.  Otherwise, if the first argument is a filename
we pass it (any any further options) to sage-run.