if any of the input class's base classes are already in the MRO of the given bases, then we end up with an irreconcilable MRO and a TypeError
this does not come up in the test cases in this module which mostly use an old-style class
with no bases as the test class (e.g. Bar).  But on Python 3 this automatically becomes a
new-style class, and this issue is exhibited because we end up trying to instantiate classes with
bases like (object, Foo), where object is already a base class of Foo.