additional enhancement to the viewcode Sphinx extension
This adds a viewcode_extra_links option which allows inserting additional
arbitrary links alongside the 'source' link that is added to class and
function documentation.  Each link can include templates for the URL,
link text, and link title, and those templates are passed context variables
related to the object being documented.

We use this in Sage to produce a link to that object's source code in the
sage GitLab project--users logged into GitLab may then edit the source
directly and submit a merge request to Sage through GitLab.