Do not use broken docker version on CircleCI
CircleCI currently uses an old docker version. Let's use the one that we used
on GitLab CI as we have done more extensive testing with that one.

Strangely, this increases the size of the resulting docker image from 611MB to
635MB on CircleCI (same as on GitLab.) I have not checked what's the difference
between the two but I guess it's safer to go with the later version.