add prototype for SafeSortable wrapper class, inspired by pprint._safe_key on Python 3
Any two objects wrapped in a SafeSortable can be ordered a la Python 2, and SafeSortable can also be used as a
sort key.

Still work to do with this on figuring out the best API, as well as performance analysis, especially on
Python 2 where it probably creates a significant performance hit, though there are things we can do to
improve that.