py3: rework this function a bit to work better for Python 3 classes
In Python 3, functions defined in a class body are still just normal
function objects, and not wrapped as unbound methods, so we need to handle
that case.  This still won't quite work if the function is passed a list
of functions as its first argument, as opposed to a class.  But I think
we can accept that for now.