Trac #27845: add Or2018a reference and cite it in two tests.
Two of the positive_operators_gens() tests in geometry/cone.py check
properties that were proved in a 2018 paper. In fact, the Sage
implementation is what led to the conjecture that the paper proves in
the first place. In any case -- now that the paper is published --
this commit adds a new global reference and cites it for those two
test cases.