Added code for random test of modular decomposition from reconstruction
Wrote functions to generate a random MD tree, and generate a graph
from a MD tree, and a test to check that the modular decompostion
of that graph matches the original tree.

Also moved the Habib-Maurer code to its own location.

Improved the gamma-class code to use itertools.chain and DisjointSet, and
to avoid sorting vertices.