This forces the source links to be on a separate line / paragraph from the class name / function name and signture itself
This is needed because some functions/classes are so long (especially if they are in a
deeply nested subpackage) that it makes some awkward layout flow issues with the default
right-floated links.

We should also see about submitting this fix upstream, perhaps as an option to the extension.
In other project's docs the default formatting for the source links may work fine.  It just
happens that Sage's docs push the limits of the default formatting.