Trac #26639: use future-proof range() iterator in geometry/cone.py
The python-2.x version of range() always returns a list. This can be
inefficient, if, for example, you're using it to sum a bunch of numbers:

  sage: sum(range(10))
  45

The code above constructs the entire list [0,1,...,9], and *then* sums
its entries. Storing the list wastes memory; an iterator that lets us
consider each element one-at-a-time (without storing the whole list)
would be more efficient.

In python-2.x, the xrange() function does what we want. But in
python-3.x, there is no xrange() -- it's been replaced by a smarter
version of range()! We don't want to leave the list-building range()
calls in there for python-2.x, but we can't "upgrade" them to xrange()
without creating problems for the python-3.x migration down the road.

So what to do? Fortunately, python-2.6 has the new version of the
range() function, but it's hidden by default. We gain access to it with

  from six.moves import range

at the top of geometry/cone.py. This automatically affects all library
code and doctests in that file, converting their uses of range() to
iterators. The code has been inspected to make sure that this is
semantically correct (that is, nothing is really expecting range() to
return a list).