Enable R error handling
This was disabled before because tests relied on it:

- a doctest in function_field used `r` as a variable name. For some
reason it didn't get shadowed properly and the doctest ended up getting
and setting invalid values. It still worked somehow, as long as `r`
didn't raise any errors.

- dumps was broken because the is_string function was not overriden in
RElement. The default always returns false. Because of that, strings
were dumped literally. To load them again, we need to dump their repr
(with quotes) instead.