some error-handling 'improvement'
Call GAP_Error_Setjmp inside a try/except before calls into GAP, so that if
an error occurs resulting in a longjmp out of GAP (e.g. from an otherwise
unhandled JUMP_TO_CATCH) we have some reasonable place for GAP to jump *to*.

As long as our custom error_handler was run a Python RuntimeError exception
should be set with the error message from GAP, and we can then catch it here.