doctest framework: add ability to specify test files that contain known failures
If a file with known failures fails, the test run is not counted as a failure
overall unless other, unknown failures occur. If the converse occurs--a file
that is known to fail succeeds--the test run is still counted as a success but
the unexpected success in that file is also reported.