copy the current environment before passing it to GAP
just passing the global environ variable is dangerous, because functions
that modify the environment such as putenv() and setenv() can cause
environ to be reallocated, and thus the original value passed by copy
to GAP is no longer valid.  In some sense this is a bug in GAP: It shouldn't
save the envp argument to InitializeGap to a variable.  But regardless
it's safest to just copy the current environment first.