fix the test for available operations on an object
(actually, conversely, this tests whether a given operation can apply
to an object)

GET_OPER_FLAGS returns a list of lists of filters (in the form of flag
lists) that are used to determine if an operation applies to an object.
For each list of filters, all of those filters should apply.