set GAP's notion of 'screen width' to the maximum allowed
This prevents GAP from inserting needless newlines when printing most GAP
objects (explicitly printed newlines will still work, just not those
automatically inserted by GAP for wrapping long lines.

Ideally there would be a way to explicitly turn off this wrapping entirely,
as it is not so useful when embedding GAP.  But this will work for the
majority of cases, including tests in Sage that currently fail due to this.