add support in sage-spkg for spkg-preinst scripts
It makes sense to support this as well as spkg-postinst; it isn't strictly as
necessary since there's not much difference between having a preinst script or
putting some steps at the beginning of spkg-build or spkg-install

However, the use case in mind for this is as a simple way to handle conflicting
packages (of which there are relatively few examples); in this case
uninstalling the gfortran package before installing gcc.  It makes sense to
keep steps like that logically separate from build/install of the package
itself.