fix pretty printing of SandpileConfig and SandpileDivisor
These classes are subclasses of dict, so they *should* be pretty-printed
the same way plain dicts are, with their elements ordered by key.  However,
IPython's pprint machinery does not take into account subclasses, so we still
must manually register them.

As noted in the comment, Sage ought to provide an easier way to do this
but we don't yet, so this is good enough.