Added the package Chow, a library for computations in intersection theory.
The (experimental) package Chow allows computations in intersection theory.
It is inspired by the Maple library "Schubert" of Sheldon Katz and Stein Arild
Strømme. It was written specifically to compute the Euler number of the
compact holomorphicially symplectic manifold of dimension eight constructed in
"Twisted cubics on cubic fourfolds" (Ch. Lehn, M. Lehn, Ch Sorger and
D. van Straten,J. Reine Angew. Math. 731 2017, 87–128), but allows all sort
of computations in intersection theory, as for example the computation of the
number of twisted cubics on general quintinc threefolds.

See the references and/or the thematic tutoriel for more details or
https://www.math.sciences.univ-nantes.fr/~sorger/chow_en.html

Even though inspired by the Schubert library, this package was written pretty
much from scratch. Whereas we are sure that the computation of the above
Euler number is correct and that we tested quite carefully the methods of this
package it is of course possible, if not likely,that some bugs remain.
If you find one, please contact the authors. In short: use at your own risk!