Reworked this a bit more
The SAGE_MP_LIBRARY variable should really only have either the values
"gmp" or "mpir" (i.e. actual package names, as opposed to "system").

When using the system GMP we just set its value to "gmp" (whether or
not it's actually MPIR).  In this case it's just a placeholder since we're
not installing either SPKG anyways.