Trac #27354: warn against chained comparisons in assume().
The python chained comparison notation "0 < x < 1" doesn't play nice
with the assume() function. The expression "0 < x < 1" is translated
by python into "0 < x and x < 1", which is not something that can be
used as an assumption. Implementation details conspire against the
user and no error is raised, but the end result is essentially junk.

This is likely hard or impossible to fix, so this commit merely
documents the behavior. A new WARNING block is added to the assume()
function, and a new example shows how to pass "0 < x < 1" as two
separate arguments. A test case is also added to ensure that the
claims made in the WARNING block remain accurate.