Trac #27354: add INPUT and OUTPUT documentation for assume().
As part of this ticket, a reviewer mentioned that it's probably not
ideal to have a big fat WARNING be the first thing that a user sees in
the documentation for the assume() function. To avoid that situation,
this commit adds some basic input and output documentation for that
function, at which point the WARNING can come after those.