Trac #27354: fix a pointless try/except in assume().
The assume() function has a try/except block that is looking for a
KeyError, but which actually throws an AttributeError (so the "except"
does nothing). Moreover, the "except" block does nothing other than
re-throw the exception, so catching it in the first place is a bit
pointless in my opinion. This commit removes the "except" block
entirely, and allows the AttributeError to propagate up. A new test
case is also added to ensure that this behavior (which will be
documented soon) remains consistent.