Trac #27354: call forget() consistently in assume() examples and tests.
When testing assumptions, it's important to call forget() because
otherwise the assumptions made in previous tests and examples are not
cleared. There are two obvious ways to do this: at the beginning of
your test, or at the end of any test that makes assumptions. Both
methods have merits; this commit takes the "clean up after yourself"
approach and moves all of the forget() calls in the assume() doctests
to the end of the block.