Trac #25680: Fix doctest worker output tempfile handling on Cygwin
This fixes the issue originally described at https://trac.sagemath.org/ticket/25107#comment:14
in a way that will work on all supported Python versions.

In Python 3.7 the tempfile module was refactored in such a way that setting
the delete=False attribute on the temporary file object no longer prevents
deletion of the file when the tempfile object goes out of scope.

Instead we explicitly use NamedTemporaryFile with delete=False, and manually
manage its deletion at the right time.