Trac #27642: Initial attempt at fix: re-run `configure` and `make all-build` after installing a new SPKG with `sage -i`.
We need to also ensure that the `SAGE_SPKG_CONFIGURE` macro detects if a dependency
is already installed as an SPKG, and skip trying to check the system for it.  In this
case the PRE and POST actions, if, provided, are still run.  A non-trivial example of
this is MPIR: If the mpir SPKG is installed we should stop checking the system for
that dependency, but we *should* still perform actions like setting SAGE_MP_LIBRARY
and SAGE_GMP_PREFIX.

This does not yet re-run `configure` with any arguments previously provided by the
developer and that may need to be fixed.