Trac #27908: py3: fix doctests in multi_polynomial_ideal
The implementation of the `random_element()` method in
`sage.rings.polynomial.multi_polynomial_ring_base` involves sets. This causes
the random elements of ideals to differ between Python 2 and 3.

The root reason for sets of polynomials to return their elements in a different
order under Python 3 seems to be the fact that the hash of polynomials
intentionally incorporates the string hash of its variable names, which is
different in Python 3. Therefore, doctests will have to account for the
possibility that the order of polynomials from sets, as well as results
computed from these, are different.

In case of the degree 2 example, no elements are chosen randomly from a
polynomial ring, but only from the base field. Thus, it is enough to set a
random seed to eliminate side effects and guarantee the same output in Python
3.