Further improvement to pickling of GapElement objects.
Adds a test demonstrating the specific case that first motivated this: Pickling
a conjugacy class of a symmetric group, which does not (yet) have a direct
equivalence in Sage implemented, but whose str() representation is the same as
the GAP code needed to reconstruct the object.  In effect, this works the same
as pickling for GapElements in the pexpect interface.  As such, it is not
guaranteed to work in all cases, but it does work in many cases.