Fixed bug concerning the combinatorial embedding of Schnyder drawings
Previously, the given embedding of the graph was ignored after the
graph was triangulated. But since the combinatorial embedding of a
planar triangulation is only unique up to the choice of the outer
face and reflection, this could lead to drawing to drawings whose
combinatorial embeddings are reflected compared to the given
combinatorial embedding.