Add strip_math_delims option to sphinxify so that its current default behavior can be disabled.
If I understand correctly, the SageNB instead uses a MathJax extension
which applies math rendering to any <span> or <div> with the "math"
class.  However, this does not work by default with the MathJax settings
in the Jupyter Notebook, so it's better to keep the math delimiters.

Also updates the code that removes them to use a regular expression
instead.