Add sphinxify_mimebundle function for sending docstrings as text+HTML "mimebundle" as understood by Jupyter.
The name and location of this code are debatable, as this is very
specifically for use with Jupyter.  For the text format we use
sage.misc.sagedoc.format which is similar to what sage_getdoc does.
Here we assume that we are being passed the original docstrings
unmodified (otherwise sphinxify won't do anything interesting).