Add the necessary default settings to enable HTML output in the Jupyter Notebook pager for the IPython shell.
We need to replace the monkeypatching of getdoc -> sage_getdoc with
getdoc -> sage_getdoc_original.  The difference is that the former also
runs sage.misc.sagedoc.format() over the docstring, whereas the latter
doesn't.  This is necessary so that by the time the docstring reaches
sphinxify_mimebundle() it hasn't been re-formatted yet.

That said, it's still not entirely clear how this change will affect
other functionality, as getdoc is used in several places by
IPython.core.oinspect.

We also need to inject our sphinxify_mimebundle() in place of the
default sphinxify in IPython.core.interactiveshell.  Unfortunate but not
easily avoidable at the moment.