Trac #28327: don't rely on coercion of base-ring zero in ModulesWithBasis.
The random_element() method of ModulesWithBasis() uses self(0) to
obtain the zero element of the module. This is usually fine, because
the default _element_constructor_() has a special case for the int 0.
However, in modules that override the element constructor and which
disable the default base-ring-to-module coercion, this leads to a
crash. Since self.zero() works just as well or better in all cases,
this commit replaces self(0) with self.zero().