Trac #28329: suppress warnings from failed swap statistics collection.
On Linux, the swap_memory() function in psutil tries to collect swap
statistics from the /proc/vmstat file. It looks for two lines named
"pswpin" and "pswpout", and throws a warning if they aren't found.
Indeed, the presence of those lines is controlled by the kernel
CONFIG_VM_EVENT_COUNTERS option, so they may not exist.

The resulting warning is superfluous because we don't use those
particular statistics for anything in SageMath. But it's rather ugly,
and is displayed to the user every time SageMath is started. And
what's worse, the additional output from commands like "sage -t"
causes some doctests to fail.

This commit wraps our two uses of swap_memory() in a catch_warnings
context manager that ignores warnings.