spkg-configure for eclib
the only version test we carry out is checking for certain header
that only appeared in 2018. We also check that we can link and run
some trivial eclib code (as this is C++, we can't do AC_SEARCH_LIBS).

We don't check for flint, as it is an optional dependence: eclib
functions without it as well.