Trac #28327: find random_element()s of trivial modules.
The random_element() method of ModulesWithBasis() puts random
coefficients in front of random basis elements. But, it doesn't check
that the basis is nonempty before it tries to do that. So in a trivial
module, it crashes instead of generating the zero element of the
module. This is easily fixed with a special case that returns zero
when the basis is empty.