SAGE_SPKG_DEPCHECK, spkg-configure for m4ri(e), givaro
macro SAGE_SPKG_DEPCHECK, to test for dependencies:
to reduce flaky boilerplate

spkg-configure for m4ri:
just call pkg-config

spkg-config for m4rie:
we do not check its version, and just rely on
that m4ri it depends upon is fresh enough (and
m4ri's version is checked).

spkg-config for givaro:
givaro has a dedicated script to test the version.