Trac #28000: Remove sagelib dependency on six
Removes all use of six within sagelib, as well as a few documentation examples
and other bits of Python 2-only code around places where six was used.

This keeps all currently known-passing tests on Python 3 passing, but is such
a large patch that it's hard to guarantee there are no small mistakes.