spkg-configure for lcalc
everybody but Debian has includes in libLfunction/
while Debian has it in Lfunction/
To deal with this, we introduce an env. var.
SAGE_LCALC_INCDIR_NOLIBPREFIX
which is set to '' in "normal case", and 'yes'
in "Debian case"

we don't check that the libLfunction has the same version as
lcalc executable (it's not obvious how to do this).
We just hope it's OK. After all, lcalc is very stable.