partially fix symbolic Γ, W, zetaderiv for ball arguments
Make these functions correclty call the implementation provided by real
or complex balls when the main argument is a ball, the implementation
exists and the argument list makes sense.

Note that symbolic functions can still return incorrect results when no
suitable dedicated implementation is available or when instructed not to
call it, as in:

    sage: b = RBF(3/2, 1e-10)
    sage: dickman_rho(b)
    0.594534891891836
    sage: zetaderiv(1, b, hold=True)
    [-3.932239737431102 +/- 4.44e-16]

This is a separate set of issues, to be fixed elsewhere.

Also not fixed by this commit is the following related bug:

    sage: RBF(zetaderiv(1, 3/2))
    [-3.932239737431102 +/- 4.44e-16]
    sage: RBF(3/2).zetaderiv(1)
    [-3.932239737431101 +/- 5.58e-16]

Finally, some implementations are still missing or too limited, leading
to things like:

    sage: gamma(RBF(1))
    1.000000000000000
    sage: gamma(RBF(3/2))
    [0.886226925452758 +/- 1.71e-16]
    sage: gamma(RBF(3/2), 1)
    TypeError: gamma() takes no arguments (1 given)