Trac #28094: update a failing doctest
The doctest was failing because the code prints more generators than
necessary to describe an ideal.  This is part of a larger problem in
the function field code, because ideals print using the fastest set of
generators available, even though a better (smaller) set might be
possible.  The print representation of an ideal might even change
after a pair of generators (with method `gens_two`) have been
computed, because the cached value now becomes the fastest set of
generators available.