fixed failing FinitePolyExtElement test
having not run the tests with --long, I hadn't noticed that the algorithm didn't work for elements in a finite field modulo a prime power. In these finite fields, adding one won't generate all elements of the field, so the search for an rth non-residue modulo q may terminate on g=0 which will not work. Instead the search now goes over the field/ring elements as generated by the iterator iter(K).