moved common code out of the algorithms and fixed doctest
as per the suggestion I moved the code that returns the answer (depending on whether the option all was chosen) out of the algorithms to avoid code duplication. I did the same for the self.is_one() case where I also exchanged the list comprehension with the multiply/append used in the not is_one() case.
Also fixed the missing doctest output.