Trac #28258: Reduce the likelihood of DLL base address collisions in temp modules created by cython()
For technical reasons specific to Cygwin, specifically for fork() to
work, it's important to give generated DLLs an "ImageBase" (the virtual
memory address at which a DLL is loaded into a process) that does not
overlap with any other DLL.  Normally Windows can relocate them if there
is an overlap, but that breaks Cygwin's fork().  This is a common issue
with Cygwin and there is more discussion at
https://trac.sagemath.org/wiki/Cygwin64Port#Rebasing

When building Sage for Cygwin we normally take care to ensure that none
of the DLLs in Sage overlap each other, but with these generated
cython() modules we can't do that.

Fortunately, GNU binutils/ld has a ready-made solution for that which
originates from the Cygwin community: A linker flag
--enable-auto-image-base, which tries to set ImageBase based on a hash
of the DLL name, such that there is low (but non-zero) probability of
collisions.  Unfortunately, there are *so* many DLLs in Sage that some
of their base addresses are already in the range assigned via
--enable-auto-image-base, so there is still a high likelihood of a
generated DLL conflicting with an existing DLL.  Fortunately, the Cygwin
community added an optional argument to --enable-auto-image-base which
allows us to set the minimum base address to one in a range that will
not conflict with Cygwin.  Unfortunately, when this was implemented,
64-bit Cygwin was still in development, and because it is a little-used
feature it was only ever supported for 32-bit binaries, and never got
carried over to 64-bit binaries (presumably not intentionally, as the
documentation suggests it should still work).

Therefore we are forced to use the simpler --image-base linker flag
which just sets a hard-coded ImageBase for the binary, but we
re-implement the "auto image base" algorithm and pass its result to
--image-base.  This requires some unnecessary code hacks, but it is in
fact quite effective, and drastically lowers the incidence of problems
caused by DLL conflicts when running tests that happen to use cython().