28819: avoid running mathematica frontend in doctests
Usually, the Mathematica interface interacts with the program `WolframKernel`.
However, a few doctests related to rendering result in the frontend program
`Mathematica` to be launched. This is because graphics cannot be
exported without access to a frontend, according to:

https://mathematica.stackexchange.com/questions/5026/generating-figures-over-remote-connection-using-terminal

These instances of `Mathematica` do not always seem to get closed
properly. When running a patchbot, this leads to a growing number of
orphan processes of `Mathematica` which do not seem to react to a
SIGTERM signal.

Marking the doctests `mathematicafrontend` effectively disables them. This
solution had already been applied by #23112 for one of the doctests in
the file.