Make docbuild merging compatible with spinx 2.1
The attributes "todo_all_todos" and "citations" were previously
implicitly initialized by sphinx. That changed in sphinx 2.1 due to some
refactoring:

https://github.com/sphinx-doc/sphinx/commit/9abb4820b18201f63d77fdccb4ef394a21442f7a
https://github.com/sphinx-doc/sphinx/commit/885d35e374b56d1d17f5036fe07b74229cdbec7f

So now we need to check for the case where they are not initialized yet.
We can't actually update to sphinx 2.1 yet, since its python3 only. This
is mostly intended to make things easier for distros and allow us to
update in the future.