Trac #29002: drop the readline dependency check for system sqlite.
Usually SAGE_SPKG_DEPCHECK ensures that we get a system package built
with the same features that the spkg would be. However, in the case of
sqlite, you have the option of building against either editline XOR
readline, and the two provide the same functionality.

The spkg builds against readline, but on macOS for example, the
default editline is used. This does no harm, and the system sqlite is
still usable for SageMath on macOS. This commit therefore drops the
SAGE_SPKG_DEPCHECK in sqlite's spkg-configure.m4, to accept a sqlite
built against editline as well.