Add spkg-configure.m4 for Python 3 to detect and use a system Python 3.
If a working system Python 3 is detected, the environment variable
SAGE_SYSTEM_PYTHON is set to the path of the Python interpreter
executable in sage-env-config.

Furthermore, if a system Python 3 is used, Sage Makefile will convert
SAGE_LOCAL into a virtualenv by using the Python 3 venv package--it
installs copies (should we change this to symlinks?) of the system
Python, and a pyvenv.cfg file to indicate to Python that SAGE_LOCAL is
an alternate virtualenv.  We use the venv module to do this rather than
the venv CLI, because this does not install the venv activation
scripts--we don't need these because sage-env already serves the same
purpose.