Trac #13940: catch invalid moduli before they crash NTL.
The ntl.zz_p() interface is not intended for production code, but it
shouldn't crash. The "context" associated with an NTL integer-mod-p
only works with p>1, and will crash SageMath if you pass it p=0 or p=1
because those calls lack sig_on() and sig_off() wrappers. Rather than
add the wrappers, this commit just adds an additional python check on
the value of the modulus when building a context. There was already a
check for modulus-too-large, so this is a natural fit.