29204: avoid breaking character art of nested elements
This changes the internal representation of character art `_breakpoints`
to keep track of the nesting of elements, so that line breaks can
preferably be inserted outside those nested elements.

This also deprecates `get_breakpoints` as its output is now less useful
due to the change of the internal representation of `_breakpoints`.