12406: move rank checks to _solve_right_nonsingular_square
This avoids checking whether the system is of full rank in
`solve_right`/`solve_left`, so that inexact rings can choose to
implement different behavior by overwriting
`_solve_right_nonsingular_square`.

For example, RDF/CDF simply attempt to solve a square system directly,
without trying to check whether it is of full rank.

If `_solve_right_nonsingular_square` chooses to raise a
`ValueError("not of full rank")`, then _solve_right_general` is used as
a fallback.

The documentation is updated accordingly.

This also adjusts a doctest in quivers/morphism.py to work over an exact
ring because the doctested method does not work reliably over inexact rings.