Trac #29369: new spkg-configure.m4 for brial.
This is a more or less standard spkg-configure.m4 file, complicated
only by the fact that BRiAl is a C++ library and provides no
pkg-config file. This check ensures that we can compile, link, and run
a program that uses libbrial and libbrial_groebner, but does not test
any specific features of those libraries.