Trac #29453: use PKG_CHECK_MODULES to find a system fflas-ffpack.
We already have an spkg-configure.m4 for fflas-ffpack, but it doesn't
detect the system copy yet. This commit adds a call to PKG_CHECK_MODULES
to find a new(ish) copy on the system that should work for distros who
ship the pkg-config file. I have left the existing processor flags code
more or less alone, noting that it doesn't get executed when the user
passes --without-system-fflas-ffpack to ./configure.