Trac #29345: fix most autotools bashisms.
Our autotools macros commit two main bashisms:

  * The use of FOO+="${BAR}" to concatenate the strings FOO and BAR; and
  * the use of $'\n' as a literal newline character.

The first is easy to address, by using FOO="${FOO} ${BAR}". The latter
is a bit more tricky, since the "obvious" way to add a newline into
the variable FOO is with FOO="... $(printf '\n')", but the
specification states that any newlines should be stripped from the end
of the subshell's output! To work around that difficulty, this commit
reworks the way a few of our variables are constructed.

Instead of appending new stuff to the variable followed by a backslash
and a newline, we now append the newline followed by some spaces and
the new stuff. The end result is more or less the same -- it's the stuff
we want separated by newlines and backslashes -- but with one important
practical difference. Since the newlines are printed at the beginning
(followed by some spaces), the output from e.g. $(printf '\n   ') is
left untouched. This looks a bit crazy when you see it, but it generates
the correct output in the Makefiles.