Trac #29345: fix one more newline-constant bashism in SAGE_SPKG_ENABLE.
This is essentially the same fix used in SAGE_SPKG_COLLECT and in our
configure.ac file. It reorders some things to print newlines followed
by the stuff we want, rather than the other way around. This prevents
the newlines from being stripped out of the subshell's output entirely.
This macro was previously using AS_VAR_SET for this purpose, but for
clarity I have reduced it to a standard shell-variable definition. The
name of the variable itself is not dynamic, and this saves us from
needing to think about m4 whitespace quoting and escaping rules.