Trac #29345: replace a bash array with something portable.
Instead of storing the output of sage-num-threads.py in a bash array,
this commit stores it as a string. There are three "components" to the
output, and to get at the first and second components, we now use the
POSIX shortest prefix/suffix parameter expansions to strip off the
stuff before/after the first/last space.