Trac #29631: disable additional unused features of the linbox SPKG.
The linbox package has some additional features that depend on other
packages and can be enabled by passing --with-<package> to its
./configure script. When writing the spkg-configure.m4 for linbox,
we need to detect the same set of optional features that the linbox
SPKG enables, which is a fairly large list by default.

Fortunately, most of those optional features that are enabled by
default are not used in sage, and can be disabled. This commit
disables them by passing --without-<package> to the linbox
./configure, and removes them from its "dependencies" file. As a
result, we won't have to check for those features in the forthcoming
spkg-configure.m4. The one optional feature that remains enabled is
--with-flint, because it gets enabled "automagically" even if you try
to disable it. Enabling it ourselves makes the build more predictable;
this way, it doesn't matter if flint happens to be installed before
linbox is.