Trac #29345: don't use sage's config.status for the lrcalc build.
The lrcalc SPKG has some outdated autotools, and we worked around that
in Trac ticket 19725 (commit 75de776ab) with the line,

  cp "$SAGE_ROOT"/config/config.* .

that is intended to import only config.guess and config.sub, judging by
the comment above it. However, it's also mistakenly copying in the sage
version of config.status, causing build problems when sage was configured
with a non-bash shell.