Trac #29672: use configurable SAGE_PALP_CLASS_EXE in palp_database.py.
The spkg-configure.m4 script for the palp package will soon detect the
name of the "class" executable that should be used at ./configure
time. To pass it through to the palp_database.py file that uses it,
this commit adds the variable to the rest of the environment stuff in
src/sage/env.py and build/pkgs/sage_conf/src/sage_conf.py.in. The default
behavior is backwards-compatible and uses the string "class-4d.x".